fix(spur-core): widen mem-per-cpu product before cast; note grp_wall not enforced
Widen num_tasks * cpus_per_task to u64 before multiplying so the per-cpu
memory product can't wrap, matching the per-node branch. Document that
grp_wall_minutes is stored/displayed but not enforced (needs a persisted,
period-resetting consumed-wall accumulator that does not exist yet).

+++ crates/spur-core/src/job.rs
@@ -544,7 +544,7 @@ pub fn effective_memory_mb(spec: &JobSpec, num_nodes: u32) -> u64 {
         .map(|mem| mem * num_nodes as u64)
         .or_else(|| {
             spec.memory_per_cpu_mb
-                .map(|mem| mem * (spec.num_tasks * spec.cpus_per_task) as u64)
+                .map(|mem| mem * spec.num_tasks as u64 * spec.cpus_per_task as u64)
         })
         .unwrap_or(0)
 }
